Abilities + loading/saving with property lists
Abilities work, however they are not completely compliant yes (repeat for example).

Also change the way loading/saving map data works, with arrays of properties that are looped, rather than a list of named properties.
This makes the code less messy, but was pretty much required for abilities either way.

The base Container class is back, but only has the API methods in it, while the saving/loading is external.